+++ external-crates/move/crates/move-cli/src/sandbox/utils/mod.rs
@@ -9,7 +9,7 @@ use move_binary_format::{
     compatibility::Compatibility,
     errors::{Location, VMError},
     file_format::{AbilitySet, CompiledModule, FunctionDefinitionIndex, SignatureToken},
-    normalized_2 as normalized, IndexKind,
+    normalized, IndexKind,
 };
 use move_bytecode_utils::Modules;
 use move_command_line_common::files::{FileHash, MOVE_COMPILED_EXTENSION};
